Rename networking into dsn
